+++ src/period.h
@@ -14,7 +14,7 @@
 #define MAX_NOTE_MOD	83
 
 double	note_to_period		(struct context_data *, int, int, double);
-int	note_to_period_mix	(int, int);
+double	note_to_period_mix	(int, int);
 int	period_to_note		(int);
 int	period_to_bend		(struct context_data *, double, int, double);
 void	c2spd_to_note		(int, int *, int *);
